AzureDeploymentCode/requirement3.txt
# Core Azure ML (keep what you have)
azureml-core==1.52.0
azureml-mlflow==1.52.0

# MLFlow - MUST match versions
mlflow==2.9.2
mlflow-skinny==2.9.2  # Changed from 3.3.1 to match mlflow

# Azure dependencies - specific compatible versions
azure-mgmt-authorization==0.40.0  # Downgraded from 3.0.0
msal==1.28.0
msal-extensions==1.0.0
azure-identity==1.13.0

# Machine learning - match your model's training version
scikit-learn==1.7.1  # Upgraded from 1.3.0 to match your model
joblib==1.3.2
numpy==1.23.5

# Utilities (keep existing compatible versions)
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.5

streamlit
apache-airflow
optuna
#os
pathlib
astro